Update Part3.R
I added some color to the initial plots just so that it's a little more lively and I added some hints in the challenges where students could easily get stumped if they do not load the appropriate packages--I don't think it is always clear why they're loading specific packages so the hints will help them make sure that they aren't receiving errors code because of another syntax error. Also to use cowplot, folks must be updated to the latest version of R or at least the versions of 4.0.0 so it will not work if they do not have the updated R language. It also seems like the updated language requires the newer version of the R Studio IDE because some of the features of the graphics engine are deprecated.